+++ src/components/app.css
@@ -1,0 +1,7 @@
+.app {
+  width: 100%;
+  max-width: 80%;
+  margin: 0 auto;
+  position: relative;
+  z-index: 99999;
+}
