+++ src/components/blocks/AppBackground/component.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import "./style.css";
+
+const AppBackground = () => {
+  return (
+    <div className='background-app'>
+      <div className='stars'></div>
+      <div className='twinkling'></div>
+      <div className='clouds'></div>
+    </div>
+  );
+};
+
+export default AppBackground;
