+++ src/components/blocks/AppBackground/index.js
@@ -1,0 +1,3 @@
+import AppBackground from "./component";
+
+export default AppBackground;
