+++ src/components/blocks/AppBackground/style.css
@@ -1,12 +1,3 @@
-.app {
-  width: 100%;
-  max-width: 80%;
-  margin: 0 auto;
-  position: relative;
-  z-index: 99999;
-}
-
-/* background */
 @keyframes move-twink-back {
   from {
     background-position: 0 0;
@@ -101,12 +92,12 @@
 }
 
 .stars {
-  background: #000 url(../../img/stars.png) repeat top center;
+  background: #000 url(../../../assets/stars.png) repeat top center;
   z-index: 0;
 }
 
 .twinkling {
-  background: transparent url(//imapo.ru/img/twinkling.png) repeat top center;
+  background: transparent url(../../../assets/twinkling.png) repeat top center;
   z-index: 1;
   -moz-animation: move-twink-back 200s linear infinite;
   -ms-animation: move-twink-back 200s linear infinite;
@@ -116,7 +107,7 @@
 }
 
 .clouds {
-  background: transparent url(//imapo.ru/img/clouds.png) repeat top center;
+  background: transparent url(../../../assets/clouds.png) repeat top center;
   z-index: 3;
   -moz-animation: move-clouds-back 200s linear infinite;
   -ms-animation: move-clouds-back 200s linear infinite;
