+++ src/components/blocks/ErrorIndicator/index.js
@@ -1,0 +1,3 @@
+import ErrorIndicator from "./component";
+
+export default ErrorIndicator;
