+++ src/components/blocks/ErrorIndicator/style.css
