+++ src/components/blocks/Header/index.js
@@ -1,0 +1,3 @@
+import Header from "./component";
+
+export default Header;
