+++ src/components/blocks/ItemList/component.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 
-import "./item-list.css";
+import "./style.css";
 
 export default class ItemList extends Component {
