+++ src/components/blocks/ItemList/index.js
@@ -1,0 +1,3 @@
+import ItemList from "./component";
+
+export default ItemList;
