+++ src/components/blocks/PersonDetails/index.js
@@ -1,0 +1,3 @@
+import PersonDetails from "./component";
+
+export default PersonDetails;
