+++ src/components/blocks/PlanetDetails/index.js
@@ -1,0 +1,3 @@
+import PlanetDetails from "./component";
+
+export default PlanetDetails;
