+++ src/components/blocks/RandomPlanet/index.js
@@ -1,0 +1,3 @@
+import RandomPlanet from "./component";
+
+export default RandomPlanet;
