+++ src/components/blocks/Spinner/component.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import "./spinner.css";
+import "./style.css";
 
 const Spinner = () => {
   return (
