+++ src/components/blocks/Spinner/index.js
@@ -1,0 +1,3 @@
+import Spinner from "./component";
+
+export default Spinner;
