+++ src/components/blocks/StarshipDetails/index.js
@@ -1,0 +1,3 @@
+import StarshipDetails from "./component";
+
+export default StarshipDetails;
