--- src/components/error-indicator/index.js
@@ -1,3 +1,0 @@
-import ErrorIndicator from "./error-indicator";
-
-export default ErrorIndicator;
