--- src/components/header/index.js
@@ -1,3 +1,0 @@
-import Header from "./header";
-
-export default Header;
