--- src/components/item-list/index.js
@@ -1,3 +1,0 @@
-import ItemList from "./item-list";
-
-export default ItemList;
