--- src/components/person-details/index.js
@@ -1,3 +1,0 @@
-import PersonDetails from "./person-details";
-
-export default PersonDetails;
