--- src/components/planet-details/index.js
@@ -1,3 +1,0 @@
-import PlanetDetails from "./planet-details";
-
-export default PlanetDetails;
