--- src/components/random-planet/index.js
@@ -1,3 +1,0 @@
-import RandomPlanet from "./random-planet";
-
-export default RandomPlanet;
