--- src/components/spinner/index.js
@@ -1,3 +1,0 @@
-import Spinner from "./spinner";
-
-export default Spinner;
