--- src/components/starship-details/index.js
@@ -1,3 +1,0 @@
-import StarshipDetails from "./starship-details";
-
-export default StarshipDetails;
